Built new repository for Module System.
Modules or plugins for .NET are DLLs which are built from Class Library
projects. The module system attempts to import these DLLs and activate
their assembly in order to load in and activate class instances if
possible. This let's you build a backend for plugging new features into
your application.

Only verified classes will be loaded though--to make sure that
unnecessary classes/DLLs are not loaded into your application if they
reside in the application's module directory.